+++ src/main/java/com/fine/bookmanagesystem/service/InventoryService.java
@@ -8,9 +8,9 @@ public interface InventoryService {
 
     List<Inventory> getAllInventory();
 
-    Inventory getInventoryByProductId(Long productId);
+    Inventory getInventoryByProductId(int productId);
 
     void updateInventory(Inventory inventory);
 
-    void adjustInventory(Long productId, int quantity);
+    void adjustInventory(int productId, int quantity);
 }
